Add filters on user queries (#812)
* add: filters for user queries
* update: filter name, add: tests
* styling: fix whitespace issues

+++ lib/WP_Auth0_UsersRepo.php
@@ -120,6 +120,18 @@ class WP_Auth0_UsersRepo {
 			return null;
 		}
 
+		/**
+		 * Short-circuits the user query below.
+		 *
+		 * Returning a WP_User object will stop the method here and use the returned user.
+		 *
+		 * @param string $id The Auth0 ID.
+		 */
+		$check = apply_filters( 'find_auth0_user', null, $id );
+		if ( $check instanceof WP_User ) {
+			return $check;
+		}
+
 		$query = [
 			// Limiting the returned number and this happens on login so some delay is acceptable.
 			// phpcs:ignore WordPress.DB.SlowDBQuery
@@ -168,6 +180,20 @@ class WP_Auth0_UsersRepo {
 	 * @since 3.8.0
 	 */
 	public static function get_meta( $user_id, $key ) {
+
+		/**
+		 * Short circuits the return value of the Auth0 user meta field.
+		 *
+		 * Returning a non null value will stop the method here and use the returned value.
+		 *
+		 * @param integer $user_id The user ID.
+		 * @param string  $key     The meta key.
+		 */
+		$check = apply_filters( 'auth0_get_meta', null, $user_id, $key );
+		if ( $check !== null ) {
+			return $check;
+		}
+
 		global $wpdb;
 		return get_user_meta( $user_id, $wpdb->prefix . $key, true );
 	}
@@ -184,6 +210,21 @@ class WP_Auth0_UsersRepo {
 	 * @since 3.11.0
 	 */
 	public static function update_meta( $user_id, $key, $value ) {
+
+		/**
+		 * Short circuits updating a user's Auth0 meta values.
+		 *
+		 * Returning a non null value will stop the method here.
+		 * The returned value is a boolean indicating whether or not the update was successful.
+		 *
+		 * @param integer $user_id The user ID.
+		 * @param string  $key     The meta key.
+		 */
+		$check = apply_filters( 'auth0_update_meta', null, $user_id, $key );
+		if ( $check !== null ) {
+			return (bool) $check;
+		}
+
 		global $wpdb;
 		return update_user_meta( $user_id, $wpdb->prefix . $key, $value );
 	}
@@ -199,6 +240,20 @@ class WP_Auth0_UsersRepo {
 	 * @since 3.11.0
 	 */
 	public static function delete_meta( $user_id, $key ) {
+
+		/**
+		 * Short circuits deleting a user's Auth0 meta values.
+		 *
+		 * Returning a non null value will stop the method here.
+		 * The returned value is a boolean indicating whether or not the deletion was successful.
+		 *
+		 * @param integer $user_id The user ID.
+		 * @param string  $key     The meta key.
+		 */
+		$check = apply_filters( 'auth0_delete_meta', null, $user_id, $key );
+		if ( $check !== null ) {
+			return (bool) $check;
+		}
 		global $wpdb;
 		return delete_user_meta( $user_id, $wpdb->prefix . $key );
 	}
